@@ -2,7 +2,7 @@ import styles from "./page.module.css";
 
 export default function Home() {
   return (
-    <main className={styles.main}>
+    <main className={`${styles.main} text-3xl font-bold underline`}>
       Hi, Khoj here.
     </main>
   );
