@@ -445,7 +445,7 @@ def get_sample_data(type):
 
 ** Installation
 *** Direct
-     - Put ~khoj.el~ in your Emacs load path. For e.g ~/.emacs.d/lisp
+     - Put ~khoj.el~ in your Emacs load path. For e.g. ~/.emacs.d/lisp
      - Load via ~use-package~ in your ~/.emacs.d/init.el or .emacs file by adding below snippet
        #+begin_src elisp
          ;; Khoj Package
@@ -467,12 +467,12 @@ def get_sample_data(type):
 ** Usage
    1. Call ~khoj~ using keybinding ~C-c s~ or ~M-x khoj~
    2. Enter Query in Natural Language
-      e.g "What is the meaning of life?" "What are my life goals?"
+      e.g. "What is the meaning of life?" "What are my life goals?"
    3. Wait for results
       *Note: It takes about 15s on a Mac M1 and a ~100K lines corpus of org-mode files*
    4. (Optional) Narrow down results further
       Include/Exclude specific words from results by adding to query
-      e.g "What is the meaning of life? -god +none"
+      e.g. "What is the meaning of life? -god +none"
 
 """,
             "readme.org": """
